WIP: Signal classes
This module is a proposal to provide classes for the different types of
signals you encounter as acoustician.

For instantaneous values of e.g. a recording of the sound pressure or
particle velocity:
* InstantaneousScalar
* InstantaneousVector

After time integration (Fast/Slow) but still on a linear scale:
* LinearScalar
* LinearVector

When those two are expressed in decibel:
* LevelScalar
* LevelVector

The last two can also be used for calculating equivalent levels.

It should be possible to easily go from instantaneous sound pressure to
say equivalent sound pressure levels, to apply calibration factors, to
plot the data, and to calculate all kinds of descriptors from the data.

Obviously, similar constructs are necessary in the frequency domain.
One or more spectrum classes will be developed which use
:class:`acoustics.signal.Frequencies` and they should also provide easy
conversion from narrowband to say 1/3-octaves and 1/1-octaves.

Also, TimeSpectrum and derivates will offer a way to work with time
series of spectral data. It will be possible to just select a spectrum
at a certain time instance (returning an instance of Spectrum), or
select a certain time frequency (returning one of the classes used for
decsribing time signals).