Fix sdist for building of wheel, fixes #243
In order to build a wheel all files in it need to have a timestamp of
January 1st, 1980 or later. In Nixpkgs we typically set all files to
January 1st, 1970, thus breaking Zip files (which wheels are). To handle
that, we explicitly set the timestamps of files that are to be used with
Zip files to 1980. That's exactly what this change does as well.